adding logic to handle input errors

+++ app.py
@@ -5,12 +5,21 @@ from lib.redfin import RedFinImages
 
 redfin = RedFinImages()
 
+def check_user_input(f_name, l_name, email, url, images):
+	if (f_name == '') or (l_name == '') or (email == ''):
+		return False
+	elif (url == '') and (images == None):
+		return False
+	elif (url != '') and (images != None):
+		return False
+	return True
+
 st.set_page_config(
-	page_title="LeadState Reel Generator :sunglasses:",
+	page_title="LeadState Reel Generator",
 	layout="wide"
 )
 
-st.title("LeadState Reel Generator")
+st.title("LeadState Reel Generator :film_projector:")
 
 with st.form(key="redfin_user_form"):
 	st.subheader("User Information")
@@ -30,14 +39,18 @@ with st.form(key="redfin_user_form"):
 	images = st.file_uploader("Upload images", accept_multiple_files=True)
 	submit = st.form_submit_button(label="Submit")
 
+
 if submit:
-	dir = redfin.download_images(url)
-	fp = open('images/{dir}.zip'.format(dir=dir), "rb")
-	st.download_button(
-		label="Download Images",
-		data=fp,
-		file_name='images.zip',
-		mime='application/zip'
-	)
-	fp.close()
-	shutil.rmtree('images')
+	if check_user_input(f_name, l_name, email, url, images) == False:
+		st.error("Please fill out all required fields")
+	else:
+		dir = redfin.download_images(url)
+		fp = open('images/{dir}.zip'.format(dir=dir), "rb")
+		st.download_button(
+			label="Download Images",
+			data=fp,
+			file_name='images.zip',
+			mime='application/zip'
+		)
+		fp.close()
+		shutil.rmtree('images')
